requirements.txt
asgiref==3.7.2
atomicwrites==1.4.1
attrs==23.1.0
certifi==2023.7.22
charset-normalizer==2.0.12
colorama==0.4.6
defusedxml==0.7.1
diff-match-patch==20230430
Django==3.2
django-filter==23.3
django-import-export==3.3.3
django-utils-six==2.0
djangorestframework==3.12.4
djangorestframework-simplejwt==5.3.0
et-xmlfile==1.1.0
idna==3.4
iniconfig==2.0.0
MarkupPy==1.14
odfpy==1.4.1
openpyxl==3.1.2
packaging==23.2
pluggy==0.13.1
py==1.11.0
PyJWT==2.1.0
pytest==6.2.4
pytest-django==4.4.0
pytest-pythonpath==0.7.3
pytz==2023.3.post1
PyYAML==6.0.1
requests==2.26.0
sqlparse==0.4.4
tablib==3.5.0
toml==0.10.2
typing_extensions==4.8.0
urllib3==1.26.18
xlrd==2.0.1
xlwt==1.3.0

